Validaciones

+++ src/views/Menu/components/Paraclinicos/components/MainParaclinicos/MainParaclinicos.js
@@ -129,35 +129,40 @@ const MainParaclinicos = props =>{
   const handleData = event =>{
     event.preventDefault();
 
-    
-    sendParaclinicos(formState.values.type, formState.values.value, formState.values.comment, 
-      localStorage.getItem('p_id'), moment().format('DD/MM/YYYY'));
-    
-    if(formState.values.type === 'trigliceridos'){
-   
-      localStorage.setItem("tri", formState.values.value)
-      localStorage.setItem("tri_C", formState.values.comment)
-    }
-    if(formState.values.type === 'glicemia'){
-      
-      localStorage.setItem("gli", formState.values.value)
-      localStorage.setItem("gli_C",  formState.values.comment)
-    }
-    if(formState.values.type === 'hemoglobina_glicosilada'){
-      
-      localStorage.setItem("hg", formState.values.value)
-      localStorage.setItem("hg_C",  formState.values.comment)
-    }
-    if(formState.values.type === 'colesterol_total'){
-     
-      localStorage.setItem("clt", formState.values.value)
-      localStorage.setItem("clt_C",  formState.values.comment)
-    }
+    if(formState.values.value === undefined){
 
-    formState.values.type = "";
-    formState.values.value = "";
-    formState.values.comment = "";
-    history.push("/menu");
+      window.confirm("Debe ingresar un valor y un tipo de paraclínico para poder ingresar valores")
+    }else {
+    
+      sendParaclinicos(formState.values.type, formState.values.value, formState.values.comment, 
+        localStorage.getItem('p_id'), moment().format('DD/MM/YYYY'));
+      
+      if(formState.values.type === 'trigliceridos'){
+    
+        localStorage.setItem("tri", formState.values.value)
+        localStorage.setItem("tri_C", formState.values.comment)
+      }
+      if(formState.values.type === 'glicemia'){
+        
+        localStorage.setItem("gli", formState.values.value)
+        localStorage.setItem("gli_C",  formState.values.comment)
+      }
+      if(formState.values.type === 'hemoglobina_glicosilada'){
+        
+        localStorage.setItem("hg", formState.values.value)
+        localStorage.setItem("hg_C",  formState.values.comment)
+      }
+      if(formState.values.type === 'colesterol_total'){
+      
+        localStorage.setItem("clt", formState.values.value)
+        localStorage.setItem("clt_C",  formState.values.comment)
+      }
+
+      formState.values.type = "";
+      formState.values.value = "";
+      formState.values.comment = "";
+      history.push("/menu");
+    }
   }
 
   return(
